+++ core/assets/assets.json
@@ -1,10 +1,6 @@
 {
   "textures": {
-    "shared:earth": {
-      "file": "shared/earthtile.png",
-      "wrapU": "repeated",
-      "wrapV": "repeated"
-    },"shared:all": {
+    "shared:all": {
       "file": "shared/alltile.png",
       "wrapU": "repeated",
       "wrapV": "repeated"
@@ -20,41 +16,8 @@
       "wrapV": "repeated"
     },
     "shared:goal": "platform/Door.png",
-    "rocket:rocket": "rocket/rocket.png",
-    "rocket:main": {
-      "file": "rocket/flames.png",
-      "sprites": {
-        "fire": {
-          "rows": 1,
-          "cols": 4
-        }
-      }
-    },
-    "rocket:left": {
-      "file": "rocket/flames-left.png",
-      "sprites": {
-        "fire": {
-          "rows": 1,
-          "cols": 4
-        }
-      }
-    },
-    "rocket:right": {
-      "file": "rocket/flames-right.png",
-      "sprites": {
-        "fire": {
-          "rows": 1,
-          "cols": 4
-        }
-      }
-    },
-    "rocket:crate01": "rocket/crate01.png",
-    "rocket:crate02": "rocket/crate02.png",
-    "platform:dude": "platform/dude.png",
-    "platform:combined": "platform/dudecombined.png",
     "platform:SomniPhobia_stand": "platform/SomniPhobiaStand.png",
     "platform:PhobiaSomni_stand": "platform/PhobiaSomniStand.png",
-    "platform:rope": "platform/ropebridge.png",
     "platform:barrier": "platform/barrier.png",
     "platform:bullet": "platform/bullet.png",
     "platform:somni_stand": "platform/SomniStand.png",
@@ -74,35 +37,15 @@
     "platform:phobia_somni_dash_side": "platform/PhobiaSomniSide.png",
     "platform:phobia_somni_dash_up": "platform/PhobiaSomniUp.png",
     "platform:background_light": "platform/BG_light.png",
-    "platform:background_dark": "platform/BG_dark.png",
-    "ragdoll:bubble": "ragdoll/bubble.png",
-    "ragdoll:torso": "ragdoll/tux_body.png",
-    "ragdoll:head": "ragdoll/ProfWhite.png",
-    "ragdoll:thigh": "ragdoll/tux_thigh.png",
-    "ragdoll:shin": "ragdoll/tux_shin.png",
-    "ragdoll:arm": "ragdoll/tux_arm.png",
-    "ragdoll:forearm": "ragdoll/tux_forearm.png",
-    "ragdoll:crosshair": "ragdoll/crosshair.png",
-    "ragdoll:foreground": "ragdoll/foreground.png",
-    "ragdoll:background": "ragdoll/background.png"
+    "platform:background_dark": "platform/BG_dark.png"
   },
     "jsons"  : {
-        "rocket:constants":  "rocket/constants.json",
-        "platform:constants": "platform/constants.json",
-        "ragdoll:constants":  "ragdoll/constants.json"
+        "platform:constants": "platform/constants.json"
     },
     "sounds" : {
-        "rocket:bump" : "rocket/bump.mp3",
-        "rocket:mainburn" : "rocket/afterburner.mp3",
-        "rocket:leftburn" : "rocket/sideburner-left.mp3",
-        "rocket:rightburn": "rocket/sideburner-right.mp3",
         "platform:jump" : "platform/jump.mp3",
         "platform:pew"  : "platform/pew.mp3",
         "platform:plop" : "platform/plop.mp3",
-        "ragdoll:glub1" : "ragdoll/bubble01.mp3",
-        "ragdoll:glub2" : "ragdoll/bubble02.mp3",
-        "ragdoll:glub3" : "ragdoll/bubble03.mp3",
-        "ragdoll:glub4" : "ragdoll/bubble04.mp3"
     },
     "fonts": {
         "shared:retro": {
