+++ core/src/edu/cornell/gdiac/somniphobia/GDXRoot.java
@@ -11,12 +11,12 @@
  * Based on original PhysicsDemo Lab by Don Holden, 2007
  * Updated asset version, 2/6/2021
  */
- package edu.cornell.gdiac.physics;
+ package edu.cornell.gdiac.somniphobia;
 
 import com.badlogic.gdx.*;
 import edu.cornell.gdiac.util.*;
 import edu.cornell.gdiac.assets.*;
-import edu.cornell.gdiac.physics.platform.*;
+import edu.cornell.gdiac.somniphobia.game.*;
 
 /**
  * Root class for a LibGDX.
