+++ core/src/edu/cornell/gdiac/somniphobia/GameCanvas.java
@@ -15,7 +15,7 @@
  * Based on original PhysicsDemo Lab by Don Holden, 2007
  * LibGDX version, 2/6/2015
  */
-package edu.cornell.gdiac.physics;
+package edu.cornell.gdiac.somniphobia;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.math.*;
