+++ core/src/edu/cornell/gdiac/somniphobia/game/DudeModel.java
@@ -8,15 +8,15 @@
  * Based on original PhysicsDemo Lab by Don Holden, 2007
  * Updated asset version, 2/6/2021
  */
-package edu.cornell.gdiac.physics.platform;
+package edu.cornell.gdiac.somniphobia.game;
 
 import com.badlogic.gdx.math.*;
 import com.badlogic.gdx.graphics.*;
 import com.badlogic.gdx.physics.box2d.*;
 
 import com.badlogic.gdx.utils.JsonValue;
-import edu.cornell.gdiac.physics.*;
-import edu.cornell.gdiac.physics.obstacle.*;
+import edu.cornell.gdiac.somniphobia.*;
+import edu.cornell.gdiac.somniphobia.obstacle.*;
 
 /**
  * Player avatar for the plaform game.
