+++ core/src/edu/cornell/gdiac/somniphobia/obstacle/ComplexObstacle.java
@@ -14,14 +14,14 @@
  * Based on original PhysicsDemo Lab by Don Holden, 2007
  * LibGDX version, 2/6/2015
  */
-package edu.cornell.gdiac.physics.obstacle;
+package edu.cornell.gdiac.somniphobia.obstacle;
 
 import com.badlogic.gdx.utils.*;
 import com.badlogic.gdx.math.*;
 import com.badlogic.gdx.physics.box2d.*;
 import com.badlogic.gdx.physics.box2d.BodyDef.BodyType;
 
-import edu.cornell.gdiac.physics.*;  // For GameCanvas
+import edu.cornell.gdiac.somniphobia.*;  // For GameCanvas
 
 /**
  * Composite model class to support collisions.
